Sorter replaced with ModuleSorter, updated class description
